requirements.txt
Flask
flask-restful
flask-cors
python-dotenv
gunicorn
psycopg2
flask-sqlalchemy
openai